breaks up json classes into separate files
