+++ rxNorm.Net.Api.Wrapper/rxNorm.Net.Api.Wrapper/Dtos/DoseFormConcept.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace rxNorm.Net.Api.Wrapper.Dtos
+{
+    public class DoseFormConcept
+    {
+        [JsonPropertyName("doseFormRxcui")]
+        public string DoseFormRxCUI { get; set; }
+
+        [JsonPropertyName("doseFormName")]
+        public string DoseFormName { get; set; }
+    }
+}
